convert doc/Makefile to CMake
CMake will build the pdf version of the manual if texi2pdf is installed.
It will be installed to DOCDIR/ledger{,3}.pdf, for example
/usr/local/share/ledger/ledger3.pdf.
Also, the man page will be installed to MANDIR/man1/ledger.1

The option BUILD_DOCS is now on by default.

A new option BUILD_WEB_DOCS is used to toggle the generation
of the html version of the manual and the man page (off by default).

All this is added to the 'doc' make target.