Rename @code{--option} to @option{--option}
Rename @code{-Q} to @option{--download (-Q)}. The reasoning is that the reader of the manual is a newcomer, and as a newcomer --download is more self-explicit compared to a cryptic single letter. But information is kept for advanced user who can still exercise their mastery with cryptic things.

Fix short option -Z (and not -L) for --price-exp

Fix "@section Balance format" (as @option in title is causing issues with makeinfo).

Fix consistency of @option{--option @var{VAR}} by adding systematically @var{VAR}.

Add period at end of sentences.