[emacs] Fix bug 1060.
Set as an option to kill (or not) the reconcile buffer after
ledger-reconcile-finish.  The default is set to *not* kill the
buffer. This is justified as the facts that:

1. In old ledger mode (the one from Ledger 2.6), the default is to *not*
kill the buffer. Also for the ones who will migrate from 2.6 to 3 when
package will be available in their linux distribution, this will be the
path of least surprise.
2. The current documentation does state that the buffer is to be killed
with 'q' after 'C-c C-c'.

This has been made configurable, as killing buffer may already be part
of the workflow of some persons.

[ci skip]