Add the option to make ledger-report-mode use the header-line.
At present `ledger-report-do' will insert the report name, command, and
a separator line of === characters into the buffer as a preface to the
output of the report command itself.

This commit adds the option of using the `header-line' to display this
information, using the `ledger-report-use-header-line' switch (nil by
default). The function evaluated in the header line can also be
customized by setting `ledger-report-header-line-fn'; the default
function prints the report name, buffer, and command.

Signed-off-by: Aaron Jacobs <atheriel@gmail.com>