use (ledger-master-file) where applicable
instead of always using (current-buffer), we should play well with those
using multiple ledger files.

For example, (ledger-display-balance-at-point) can rely on
(ledger-master-file) to return the correct buffer to calculate balance
for account at point.

The reason for this is that if (current-buffer) doesn't contain all
ledger data, (ledger-display-balance-at-point) will return the wrong
balance for the account.

Signed-off-by: Felipe Balbi <balbi@kernel.org>