Keep test/ directory of the load-path
Because the files, features, and symbols in "test/" are not prefixed
with "ledger-", we should make it as unlikely as possible for "test/"
to end up on the `load-path'.

By adding the file "test/.nosearch" at least
`normal-top-level-add-to-load-path' and the tools used to maintain the
Emacsmirror are told to ignore test.  Homegrown package managers might
also benefit.

In "test/Makefile", explicitly add "." to the `load-path' to counter
this change in the one case where we want these files to be on the
`load-path'.