Add support for flymake in Emacs 26+
Flymake, Emacs's built-in on-the-fly syntax checker, got a major
overhaul in Emacs 26. This means we can get built-in syntax checking
without requiring users to install both flycheck.el and
flycheck-ledger.el. The checking is done via `ledger balance`. Things
to keep in mind that as of now (the initial implementation):

- It will honor settings from users ~/.ledgerrc files (I'm not sure if
  this is a good thing or a bad thing)

- It will error if Emacs can't find ledger (i.e. if (executable-find
  "ledger") fails)

- There is no way yet to customize calls (using --pedantic, for
  example), though this would be easy to implement in the future