travis: allow emacs-snapshot/ledger-next combination to fail
Currently Travis is on Ubuntu 14.04 Trusty, and the Ubuntu-Elisp PPA
only build emacs-snapshots through 2016. This causes the ledger-mode
tests to fail because the Emacs version is 26 but they don't have the
flymake rewrite. This shouldn't be a problem in reality because people
who are building Emacs from the git branch have almost certainly
updated their builds since the flymake rewrite in 2017.