Add support for flymake in Emacs 26+ (#87)
* Add support for flymake in Emacs 26+

Flymake, Emacs's built-in on-the-fly syntax checker, got a major
overhaul in Emacs 26. This means we can get built-in syntax checking
without requiring users to install both flycheck.el and
flycheck-ledger.el. The checking is done via `ledger balance`. Things
to keep in mind that as of now (the initial implementation):

- It will honor settings from users ~/.ledgerrc files (I'm not sure if
  this is a good thing or a bad thing)

- It will error if Emacs can't find ledger (i.e. if (executable-find
  "ledger") fails)

- There is no way yet to customize calls (using --pedantic, for
  example), though this would be easy to implement in the future

* Mention in the readme flymake support

* Minor cleanup

* Try not to enable flymake if flycheck-ledger is available

* ledger-flymake: prefer 'ledger-binary-path' over calls to ledger

* ledger-flymake: satisfy checkdoc & byte compiler

* Fix wording in README

* flymake: improve loading

* flymake: add defcustoms for --pedantic/explicit

* flymake: comment changes

Prefer double space after sentences

* Fix backquote style

* Fix ledger-flymake requires

* ledger-flymake: doc fix

But checkdoc gets angry now :-(

* Check for supported Emacs version before trying ledger-flymake

* ledger-flymake: require cl-lib

* ledger-flymake: Don't auto-enable

- Add an ;;;###autoload so that users can enable if they want.

- Remove the require statement from ledger-mode.el

- renamed the function that enables flymake so that everything starts
  with ledger-flymake-<etc>

* Silence byte compiler

* ledger-flymake: add require flymake statement

* travis: allow emacs-snapshot/ledger-next combination to fail

Currently Travis is on Ubuntu 14.04 Trusty, and the Ubuntu-Elisp PPA
only build emacs-snapshots through 2016. This causes the ledger-mode
tests to fail because the Emacs version is 26 but they don't have the
flymake rewrite. This shouldn't be a problem in reality because people
who are building Emacs from the git branch have almost certainly
updated their builds since the flymake rewrite in 2017.