Use `regexp-opt-depth` to count capture groups
This fixes some bugs in the manual group-counting, which in particular made
incorrect assumptions about the number of groups in regexps not containing
sub-regexps.

It is safe to make this change to the group indices, because the only regexp
whose group indices change is `ledger-xact-line-regexp`, which doesn't appear to
be used anywhere.