Fix group indexes of repeated sub-regexps
We normally don't increment the group index when we encounter the same regexp,
because we are probably accessing a different subgroup of it.  But for the case
of full-date, we actually match two separate occurrences of iso-date, so we need
to indicate somehow that the group index should be incremented.