feat(xact): add default time when copying transaction
Use the transaction's time as default time when asking when the copy should be
created.  This allows us to use the relative syntax of org-read-date to move the
transaction an <interval> forward or backward from the time of the transaction:

+1d (+ 1 day from TODAY)
++1d (+ 1 day from DEFAULT-TIME, i.e. date of the original transaction)

In other words, single +/- is relative against today and double ++/-- is
relative against the DEFAULT-TIME.

Note that since currently the DEFAULT-TIME is today, this feature is only
strictly adding functionality and should therefore be fully backward-compatible.