Allow ledger-reconcile-insert-effective-date to be a function
Allow the user to customize ledger-reconcile-insert-effective-date by making it
a function that is called with the point at the posting to be cleared.  This
allows the function to decide whether to insert an effective date based on,
e.g., the payee or the account in question.

This is kept backwards-compatible with existing boolean values, which just
behave as a constant-returning function.