Do not visit ~/.ledgerrc in a buffer
Visiting the file may cause ledger-init-parse-initialization to attempt to parse
the unsaved contents of an existing buffer visiting that file, which would not
match the behavior of the ledger command-line program, and which is potentially
confusing.

Also, visiting the file puts the file name in recentf and other such lists,
which is probably unintended.