ledger-display-balance-at-point: Use buffer if too long
`ledger cleared` can produce output with quite a few lines, especially for
accounts with subaccounts or multiple commodities.  display-message-or-buffer
will show output in a buffer if it is too long to fit in the echo area (before
this commit, only the last few lines of the output would be displayed at all).