ledger-report: Do not erase buffer until report command is expanded
If the report contains format specifiers that cause the user to be prompted, it
is nicer to keep the currently-displayed report (if any) on screen, in case the
user wants to refer to that information while answering the prompts.

This may also make it possible, in the future, for format specifiers to
calculate defaults based on the buffer contents, such as the account name at
point.