Allow `ledger-insert-effective-date` to work with regions.
It updates or removes the effective date of all transactions that start
in the region (technically, that start on a line in the region),
without editing any postings.

With my credit card, I use the statement date as the effective date.
That means I have a long block of transactions that all have the same
effective date. This makes it easy to set them all at once.