ledger-report-expand-format-specifiers: Fix unexpanded specifiers bug
When looping over the string, using the end of the last match as the starting
point for the next search is incorrect (too far forward) if the last replacement
text was shorter than the format specifier it was expanded from.

Rather than trying to calculate the string indexes in a fiddly way, just do the
expansion in a temp buffer instead.  `replace-match` has an intuitive behavior
for this use-case, and this also has the nice benefit of avoiding some
unnecessary string copying (not that it likely makes much difference).

Fix #424.