ledger-complete: Remove ledger-parse-arguments and inline its callsites
This function is overly complex because it used to support pcomplete, but is no
longer needed for that purposes.  Only the caar of its return value was used at
either callsite.  That element is simply the part of the account or payee name
at point, upto point itself.

While we're at it, clean up some of ledger-fully-complete-xact.